Simulations/requirements.txt
joblib==0.14.1
tqdm
pandas==0.25.3
psutil==5.6.7
dask==2.10.1
matplotlib==3.1.2
numpy==1.18.1
scipy==1.4.1
numba==0.48.0
pytest==5.3.5
google-cloud-storage==1.26.0